Complete rewrite of the Tokenizer
Completely rewrote the tokenizer code imported from OmniCpp, fixing a
number of small bugs on the way; we can now add additional token types
in future releases just by modifying the s:TypeRegexMap variable.

Imported a utility function from OmniCpp to check if the cursor is
inside a comment or a string.