Add type lookup functions
Refactored the regex matching functions in local and global scopes into
scope.vim. MatchLocal() and MatchGlobal() now take an optional argument
that breaks the search when the first match is found; that way they can
be used for variable type lookup.

Added type lookup wrappers in type.vim; we still need to add the regexes
for matching variables declarations.