Rewrote the scope regex lookup routines
The limitations of a single regex became obvious when we needed to
extract multiple sub-patterns at one match location (namely type
modifiers in a declaration). We now use a master regex to detect the
pattern's location, and multiple sub-regexes to extract needed
information (the old behaviour can still be obtained by using ['.*'] for
sub regexes). This comes at a slight overhead though.