New type lookup routine for variables
The GetType() routine searches for the type of a variable using the same
precedence rules as the compiler. The auxiliary TagSearchType method,
which looks up a fully qualified variable name in tag files, is still a
stub though.

GetGlobalType() has been removed, since we can use tag files for global
variables.