Minimalist test framework
Unit tests are named b:Test*. The tests in a single file make up a
suite, which can be run by opening the file and typing :RunTests.
(The in.vim file must have been sourced before though)

The b:Assert() function is available for unit tests. Any false assertion
in a test causes the test to fail, otherwise the test is successful.