Test framework fixes
We now open the data file, defined by the b:testFile variable in the
test suite, and run all tests on that buffer. The cursor position is
determined by looking up comments of the format "//*TestName*".

Since we switch buffers, we define test functions as global. It is
possible to define them as script-local, which would avoid name clashes;
but this might take some work to implement.