ExtractInstruction() and SearchPairBack()
ExtractInstruction() searches backward for an instruction-delimiting
character, and extracts the code up to that point. It sometimes matches
more than a single instruction, but shouldn't match less.

SearchPairBack() is a string-version of the searchpair() Vim method. It
is intended to search backwards for the first opening item that
encompasses the end of the string. We expect a sanitized string, and
don't check for comments and the like.