Inherited contexts lookup function
When resolving nested contexts in ns#CurrentContexts(), we look up
inherited contexts for classes/structs recursively.

Caveats:
========
- Nested contexts, using-directives and using-declarations are
considered separately, and are not combined; therefor the following will
fail:
    namespace n1 {
        class c1 {
            void f();
        };
    }
    using namespace n1;
    void c1::f() {
        //doesn't search n1::c1...
    }
- Context precedence is not enforced; the order in which contexts are
considered is somewhat random, and the search stops after a match is
found.
- We use the same lists of includes, global declarations and global
directives for all searches; therefor, in an included file, we could end
up matching against a using-instruction that is either not visible, or
occuring after the declaration we are looking up (the afore-mentioned
lists should be recalculated for every file, but this would be too
slow).