Add file parsing routines to ns.vim
In addition to recursively parsing the current buffer up to the cursor's
position, we can now parse a list of files for using-instructions. The
local/global separation has been removed, since it is currently not
needed.

Using this, we now rebuild the list of visible using-instructions for
every matched class in ns#BaseClasses().