Partial file parsing routines in ns.vim
We now have Parse*Single() methods to parse a single file up to a
certain line number for using-instructions; that way only instructions
prior to a declaration are matched in BaseClasses(). We still don't
distinguish local/global instructions when grepping, but that should be
a minor problem.

Note: we don't add partial parses to the cache to keep cache management
simple.