Add graph.vim
A graph object is a generic structure modeling dependencies between
nodes, with a method to walk through the nodes:
- First search for children
- If none, search for siblings at every level along the inheritance tree
up to the root node.

Nodes with children are #include statements, the children being the data
in the included file. The afore-mentioned order respects #include
dependencies and visibility by first searching the included file, then
statements following that #include in the parent file and its parents,
up to the main file.

Nodes are simple objects holding a data content and references to parent
and children.