Recursively parse files
parse#Recursive() builds the parsing graph of a file, starting at the
root node, parsing new includes as they are reached with next() and
updating the graph. When we reach the end of the graph, the list of
visited nodes (with #include duplicates filtered) is returned, as it
holds all the parsed data in correct order.

Also move utils#Grep() -> parse#Grep()