Partly visible files in parse#Recursive()
We added a 'path' attribute to nodes, which retraces the path through
includes up to that node. Using this, we can determine what files are
partly visible from that node's location. Furthermore, when looking up
subsequent nodes through graph#Next(), we skip includes that were
already visited (through the filter() method) instead of filtering the
results afterwards.

The parse#Recursive() function has been updated to return both the list
of using-instructions with their visible files, and the full list of
includes reachable from the parsed file. We do not resolve
using-instructions at this point to avoid useless overhead when only the
list of includes is needed.