Update using-instructions resolution algorithm
Every using-instruction is resolved as soon as it is extracted from the
graph, and in turn the graph's current node is updated with the result;
only instructions that were resolved correctly are visible from
subsequent nodes. That way, we can assume that all previous instructions
are fully resolved when doing a tag query, and don't need to combine
them, speeding up the lookup.